supabase-support-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-telegram-bot==21.5
google-generativeai>=0.8.3
supabase==2.9.1
httpx==0.27.0
python-dotenv==1.0.1
